permissions, pdf securing, public files, other fixes

+++ app/ADMdeleteCourse/route.tsx
@@ -4,8 +4,8 @@ import { dumpObject, systemAction, systemActionStatus } from "@/functions/loggin
 import { systemLog } from "@/functions/logging/log";
 import { deleteCourse, getCourse } from "@/functions/queries/course";
 import validateSession from "@/functions/validateSession";
-import { rm001000, rm001001, rm011000, rm011002 } from "@/responses/messages";
-import { badRequest, notFound, unauthorized, unprocessableContent } from "@/responses/responses";
+import { rm001000, rm001001, rm011000, rm011002, rm011012 } from "@/responses/messages";
+import { badRequest, notFound, serviceUnavailable, unauthorized, unprocessableContent } from "@/responses/responses";
 import { NextResponse } from "next/server";
 
 export async function DELETE(req: Request, res: Response){
@@ -18,6 +18,8 @@ export async function DELETE(req: Request, res: Response){
     if (!validatedUser) { return unauthorized(rm001000) }
     const course = await getCourse(db, courseID)
     if (!course) { systemLog(systemAction.ADMdeleteCourse, systemActionStatus.error, rm011000, validatedUser, db); return notFound(rm011000) }
+    if (!course.available) { systemLog(systemAction.ADMeditCourse, systemActionStatus.error, rm011012, validatedUser, db); return serviceUnavailable(rm011012) }
+    if (course.permissionRequired > validatedUser.status) { systemLog(systemAction.ADMdeleteCourse, systemActionStatus.error, rm001000, validatedUser, db); return unauthorized(rm001000) }
     const signups = await getCourseSignupsCount(db, courseID)
     if (signups != 0) { systemLog(systemAction.ADMdeleteCourse, systemActionStatus.error, rm011002, validatedUser, db); return unprocessableContent(rm011002) }
     await deleteCourse(db, courseID)
